app.py
import streamlit as st
from dotenv import load_dotenv

# from langchain.memory import ConversationBufferMemory
from langchain.chains import ConversationalRetrievalChain
from langchain.llms import HuggingFaceHub


def main():
    load_dotenv()


if __name__ == "__main__":
    main()
